Fix ANRs that happened in xDrip.
The main issue fixed here is that fact that on a first call of LibreTrendUtil.getData the value of m_libreTrendLatest.timestamp might be zero. In this case the data that will be read from the DB will be the entire data.
Depending on how big your DB is, it might take a long time, and creates an ANR.
After that, xdrip might also remove LibreTrendUtil from memory.

The fix is simple: only read the last 60 minutes if nothing else was specified.